清除索引目录时忽略写锁文件

+++ tnxjee/tnxjee-repo-lucene/src/main/java/org/truenewx/tnxjee/repo/lucene/support/LuceneIndexRepoSupport.java
@@ -443,7 +443,9 @@ public abstract class LuceneIndexRepoSupport<T> implements IndexRepo<T> {
         try {
             String[] names = directory.listAll();
             for (String name : names) {
-                directory.deleteFile(name);
+                if (!"write.lock".equals(name)) {
+                    directory.deleteFile(name);
+                }
             }
             writer.close();
         } catch (IOException e) {
